SQL_PROJECT-ANALYZE-INT-DEBTS
Using SQL to analyze World Bank debt given to countries CREDITS: This project is done under the 'Datacamp' URL:https://projects.datacamp.com/projects/754 .

The World Bank's international debt data¶ It's not that we humans only take debts to manage our necessities. A country may also take debt to manage its economy. For example, infrastructure spending is one costly ingredient required for a country's citizens to lead comfortable lives. The World Bank is the organization that provides debt to countries. In this notebook, we are going to analyze international debt data collected by The World Bank. The dataset contains information about the amount of debt (in USD) owed by developing countries across several categories. We are going to find the answers to questions like:

What is the total amount of debt that is owed by the countries listed in the dataset? Which country owns the maximum amount of debt and what does that amount look like? What is the average amount of debt owed by countries across different debt indicators